fsm_dp: change memory model from coherent memory to pages
Change fsm_dp memory pool memory from coherent memory to up to 256K pages
using alloc_pages(). The memory pool address space is not contiguous any more.

Kernel virtual memory for the memory pool is managed as follows:
    - A cluster of 256KB is allocated using alloc_pages().
    - 256 clusters making up 64MB can be allocated for a memory pool.
    - Fixed size buffers are assigned from the clusters.
    - The pool memory address is only contiguous within a cluster
      in the kernel address space.

A buffer is contiguous in a cluster. If a buffer doesn't fit into a cluster,
the next cluster is used. The buffer is preceded with a control block of L1
cache line size. User space follows the control block. The user space of a
buffer starts with a common fsm_dp_msghdr struct followed by user data.
The user buffer size, (struct fsm_dp_mmap_cfg *) hdl->bufsz is the size of
the user buffer, which does not include the control block. When budgeting a
memory pool, other than buffer size and number of buffers, control block size,
and unused space within a cluster should be considered.

Application space mmaps clusters into contiguous application space. There is
no gap of applicaton address space from one cluster to the next cluster.  At
the end of a cluster, there is unused space, same as kernel.

Calculation to a buffer address from the buffer index needs to account for
unused space at the end of a cluster.

Change-Id: Ib8bdb48b45ae9ddb7b46a0d0b73cb3446e626469
CRs-Fixed: 2660112
Signed-off-by: Jie Luo <jluo@codeaurora.org>