fsm-oru-forwarder: FSM ORU forwarder support
Implementation of the FSM ORU Forwarder. The fsm-dp driver was
enhanced with an internal kernel API to support the ORU Forwarder.

The forwarding operation  is copyless. It is an an agent to
the fsm-dp driver. The forwarder captures a specific ethernet eth-type
traffic from a specified ethernet interface.

Upon receipt of an ethernet frame in a skb, it forwards the skb to the
FSM through fsm-dp interface.  Upon completion of transmission to the
FSM, skb is free.

If a message received from the FSM is aggregated, the Forwarder will
de-aggregate the message based on the iob list in the aggregated
header into multiple individual messages. The Forwarder creates a skb
from each message data without copying the data and then transmits
each message with an ethernet header with a specified eth-type out a
specified ethernet interface. A reference count is maintained. When
the skb is freed by the ethernet driver, the skb destructor is invoked
and the reference count is decremented. If the reference count is 0,
the fsm-dp message buffer is freed to the fsm-dp UL memory pool.

Forwarder statistics are maintained. A debugfs node is supported
for accessing forwarder statistics and to control enabling/disabling
the Forwarder.

A netlink interface is provided. It provides configuration,
statistics, and control function.

Change-Id: Ia552c474ba1c58fe781b3ffa2313d3957c0bbc30
CRs-Fixed: 2687701
Signed-off-by: Jie Luo <jluo@codeaurora.org>