irqchip/gic-v3-its: Change initialization ordering for LPIs
We currently initialize the LPIs (and the ITS) fairly early, even
before the SMP support and the CPU interface. This is a bit odd
(as LPIs are not exactly crutial for the early boot process),
and is going to cause issues when reorganizing the probing code.

Let's move this initialization later.

Change-Id: I2bfd519afa0d9dc92321b54376cf752684018654
CRs-Fixed: 2708872
Reviewed-by: Julien Thierry <julien.thierry@arm.com>
Signed-off-by: Marc Zyngier <marc.zyngier@arm.com>
Signed-off-by: Hou Zhiqiang <Zhiqiang.Hou@nxp.com>
(cherry picked from commit a3dcf288bc0275656317ffb837a81f7177721691)
Signed-off-by: Siva Reddy <sivared@codeaurora.org>