irqchip/gic-v3-its: Allow use of pre-programmed LPI tables
In order to cope with kexec and GICv3, let's try and spot when
we're booting with LPIs already enabled, and the tables already
programmed into the redistributors.

This code is currently guarded by a predicate that is always false,
meaning this is not functionnal just yet.

CRs-Fixed: 2708872
Reviewed-by: Julien Thierry <julien.thierry@arm.com>
Signed-off-by: Marc Zyngier <marc.zyngier@arm.com>
Signed-off-by: Hou Zhiqiang <Zhiqiang.Hou@nxp.com>
(cherry picked from commit 11e0c43a3ae82382d7b21abcee160b95c122d7d5)
Signed-off-by: Siva Reddy <sivared@codeaurora.org>

Change-Id: I0e8b73e2a4a944dc4221b7b901f0631c7cff097c