fsm-oru-forwarder: Fix to avoid dropping packets with vlan configured
Fix for data pass issue with vlan configuration in the forwarder.
For example, eth1.0, a vlan-device over ethernet network device eth1
is created for ORU traffic. Forwarder does dev_add_pack() to setup
callback function to intercept ORU packets from network. On the
receive callback from network stack, the second arguement, is the
netdev for the interface where packet comes from which is for eth1.0
and the forth argument is the netdev for eth1. Driver should check the
2nd argument against its configured information.

Change-Id: I486ed617a9c9a1a5333ee4262679b12d9116baa2
CRs-Fixed: 3029694
Signed-off-by: Che-Min Hsieh <cheminh@qti.qualcomm.com>
Signed-off-by: Ning Wan <nwan@codeaurora.org>